[scm] created scm module in algorithms
